requirements.txt
absl-py==2.3.1
asgiref==3.9.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.1
keras==3.10.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.1
namex==0.1.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
pyOpenSSL==25.1.0
redis==6.2.0
requests==2.32.4
rich==14.1.0
scikit-learn==1.7.1
scipy==1.15.3
service-identity==24.2.0
six==1.17.0
sklearn==0.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
zope.interface==7.2
